Create README.md
PROBLEM STATEMENT
Developing machine learning model to accurately predict early-stage Alzheimer's disease based on multimodal neuroimaging, with the goal of enabling timely diagnosis and intervention for at-risk individuals.

PROPOSED METHODOLOGY
The process of System is to be carried out in the following sequence to achieve the Required Objectives.
Image acquisition
Input of MRI image database for pre-processing data.
Preprocessing
The obtained image is then pre processed for improvement that surpasses unwanted distortions and enhances some features important for further processing. Also, the image which is given as the input may not be of standard size. So it is necessary to obtain the required image size. Smoothing will also perform for better edge detection
Segmentation
In this phase segmentation will perform using Genetic Algorithm to find out Region of Interest (ROI).
Feature Extraction
In this process an initial set of raw variables is reduced to more manageable features for processing, while still accurately and completely describing the original data set of acne images.
Classification
Feature will classify for developing Training & Testing System. So system will be evaluate on selected parameter.